+++ src/matrix.rs
@@ -1,7 +1,16 @@
 // Module declarations.
+
+#[cfg(feature = "faer")]
 pub(crate) mod faer_mat;
+
 pub(crate) mod mat;
 pub(crate) mod matrix_trait;
+
+#[cfg(feature = "nalgebra")]
 pub(crate) mod nalgebra_dmatrix;
+
+#[cfg(feature = "nalgebra")]
 pub(crate) mod nalgebra_smatrix;
+
+#[cfg(feature = "ndarray")]
 pub(crate) mod ndarray_array2;