+++ tests/vector_index_mut.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "faer")]
+use faer::Col;
+
 use linalg_traits::Vector;
 
 #[cfg(feature = "nalgebra")]
@@ -6,91 +9,95 @@ use nalgebra::{DVector, Vector3};
 #[cfg(feature = "ndarray")]
 use ndarray::Array1;
 
-#[cfg(feature = "faer")]
-use faer::Mat as FMat;
+use numtest::*;
 
 // Test conditions.
 static X: &[f64; 3] = &[1.0, 2.0, 3.0];
 
 #[test]
 fn test_vec() {
-    let x = Vec::from_slice(X);
-    assert_eq!(x.vget(0), 1.0);
-    assert_eq!(x.vget(1), 2.0);
-    assert_eq!(x.vget(2), 3.0);
+    let mut x = Vec::new_with_length(3);
+    x[0] = 1.0;
+    x[1] = 2.0;
+    x[2] = 3.0;
+    assert_arrays_equal!(x, X);
 }
 
 #[test]
 #[should_panic(expected = "index out of bounds: the len is 3 but the index is 3")]
 fn test_vec_out_of_bounds() {
-    let x = Vec::from_slice(X);
-    x.vget(3);
+    let mut x = Vec::new_with_length(3);
+    x[3] = 4.0;
 }
 
 #[test]
 #[cfg(feature = "nalgebra")]
 fn test_nalgebra_dvector() {
-    let x = DVector::from_slice(X);
-    assert_eq!(x.vget(0), 1.0);
-    assert_eq!(x.vget(1), 2.0);
-    assert_eq!(x.vget(2), 3.0);
+    let mut x = DVector::new_with_length(3);
+    x[0] = 1.0;
+    x[1] = 2.0;
+    x[2] = 3.0;
+    assert_arrays_equal!(x, X);
 }
 
 #[test]
 #[should_panic(expected = "Matrix index out of bounds.")]
 #[cfg(feature = "nalgebra")]
 fn test_nalgebra_dvector_out_of_bounds() {
-    let x = DVector::from_slice(X);
-    x.vget(3);
+    let mut x = DVector::new_with_length(3);
+    x[3] = 4.0;
 }
 
 #[test]
 #[cfg(feature = "nalgebra")]
 fn test_nalgebra_svector() {
-    let x = Vector3::from_slice(X);
-    assert_eq!(x.vget(0), 1.0);
-    assert_eq!(x.vget(1), 2.0);
-    assert_eq!(x.vget(2), 3.0);
+    let mut x = Vector3::new_with_length(3);
+    x[0] = 1.0;
+    x[1] = 2.0;
+    x[2] = 3.0;
+    assert_arrays_equal!(x, X);
 }
 
 #[test]
 #[should_panic(expected = "Matrix index out of bounds.")]
 #[cfg(feature = "nalgebra")]
 fn test_nalgebra_svector_out_of_bounds() {
-    let x = Vector3::from_slice(X);
-    x.vget(3);
+    let mut x = Vector3::new_with_length(3);
+    x[3] = 4.0;
 }
 
 #[test]
 #[cfg(feature = "ndarray")]
 fn test_ndarray_array1() {
-    let x = Array1::from_slice(X);
-    assert_eq!(x.vget(0), 1.0);
-    assert_eq!(x.vget(1), 2.0);
-    assert_eq!(x.vget(2), 3.0);
+    let mut x = Array1::new_with_length(3);
+    x[0] = 1.0;
+    x[1] = 2.0;
+    x[2] = 3.0;
+    assert_arrays_equal!(x, X);
 }
 
 #[test]
 #[should_panic(expected = "ndarray: index 3 is out of bounds for array of shape [3]")]
 #[cfg(feature = "ndarray")]
 fn test_ndarray_array1_out_of_bounds() {
-    let x = Array1::from_slice(X);
-    x.vget(3);
+    let mut x = Array1::new_with_length(3);
+    x[3] = 4.0;
 }
 
 #[test]
 #[cfg(feature = "faer")]
 fn test_faer_mat() {
-    let x = FMat::from_slice(X);
-    assert_eq!(x.vget(0), 1.0);
-    assert_eq!(x.vget(1), 2.0);
-    assert_eq!(x.vget(2), 3.0);
+    let mut x = Col::new_with_length(3);
+    x[0] = 1.0;
+    x[1] = 2.0;
+    x[2] = 3.0;
+    assert_arrays_equal!(x.as_slice(), X);
 }
 
 #[test]
 #[should_panic(expected = "Assertion failed: row < self.nrows()")]
 #[cfg(feature = "faer")]
 fn test_faer_mat_out_of_bounds() {
-    let x = FMat::from_slice(X);
-    x.vget(3);
+    let mut x = Col::new_with_length(3);
+    x[3] = 4.0;
 }