+++ tests/vector_len.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "faer")]
+use faer::Col;
+
 use linalg_traits::Vector;
 
 #[cfg(feature = "nalgebra")]
@@ -5,9 +8,6 @@ use nalgebra::{DVector, SVector};
 
 #[cfg(feature = "ndarray")]
 use ndarray::Array1;
-
-#[cfg(feature = "faer")]
-use faer::Mat as FMat;
 
 #[test]
 fn test_len_vec() {
@@ -38,5 +38,5 @@ fn test_len_ndarray_array1() {
 #[test]
 #[cfg(feature = "faer")]
 fn test_len_faer_mat() {
-    assert_eq!(FMat::<f64>::new_with_length(3).len(), 3);
+    assert_eq!(Col::<f64>::new_with_length(3).len(), 3);
 }